Optimize first stage for development
The technique above makes development a bit easier for iterations that change only the code, but not the dependencies of the application, by caching all dependencies in a layer that is separate from the source code.

Since code changes happen more often than dependency changes, this tends to make things go a little faster. The added space cost of the additional layer is restricted to the first stage, so there is really no penalty.